+++ tests/common_tests/helpers/activation_mp_hw_model.py
@@ -12,18 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-from keras.engine.input_layer import InputLayer
-
 import model_compression_toolkit as mct
-import tensorflow as tf
-
-if tf.__version__ < "2.6":
-    from tensorflow.keras.layers import Conv2D, DepthwiseConv2D, Dense, Conv2DTranspose, Reshape, ZeroPadding2D, Dropout, \
-        MaxPooling2D, Activation, ReLU, Add, PReLU, Flatten, Cropping2D, BatchNormalization
-else:
-    from keras.layers import Conv2D, DepthwiseConv2D, Dense, Conv2DTranspose, Reshape, ZeroPadding2D, \
-    Dropout, MaxPooling2D, Activation, ReLU, Add, PReLU, Flatten, Cropping2D, BatchNormalization
-
 
 hwm = mct.hardware_representation
 
@@ -55,45 +44,3 @@ def generate_hw_model_with_activation_mp(base_cfg, mp_bitwidth_candidates_list, 
         hwm.OperatorsSet("Activation", mixed_precision_configuration_options)
 
     return generated_hwm
-
-
-def generate_activation_mp_fhw_model_keras(hardware_model, name="activation_mp_keras_hwm"):
-
-    fhwm_keras = hwm.FrameworkHardwareModel(hardware_model,
-                                            name=name)
-    with fhwm_keras:
-        hwm.OperationsSetToLayers("NoQuantization", [Reshape,
-                                                     tf.reshape,
-                                                     Flatten,
-                                                     Cropping2D,
-                                                     ZeroPadding2D,
-                                                     Dropout,
-                                                     MaxPooling2D,
-                                                     tf.split,
-                                                     tf.quantization.fake_quant_with_min_max_vars])
-
-        hwm.OperationsSetToLayers("Weights_n_Activation", [Conv2D,
-                                                           DepthwiseConv2D,
-                                                           tf.nn.conv2d,
-                                                           tf.nn.depthwise_conv2d,
-                                                           Dense,
-                                                           Conv2DTranspose,
-                                                           tf.nn.conv2d_transpose])
-
-        hwm.OperationsSetToLayers("Activation", [tf.nn.relu,
-                                                 tf.nn.relu6,
-                                                 hwm.LayerFilterParams(ReLU, negative_slope=0.0),
-                                                 hwm.LayerFilterParams(Activation, activation="relu"),
-                                                 tf.add,
-                                                 Add,
-                                                 PReLU,
-                                                 tf.nn.swish,
-                                                 hwm.LayerFilterParams(Activation, activation="swish"),
-                                                 tf.nn.sigmoid,
-                                                 hwm.LayerFilterParams(Activation, activation="sigmoid"),
-                                                 tf.nn.tanh,
-                                                 hwm.LayerFilterParams(Activation, activation="tanh"),
-                                                 InputLayer,
-                                                 BatchNormalization])
-
-    return fhwm_keras
