+++ tests/keras_tests/feature_networks_tests/feature_networks/bias_correction_dw_test.py
@@ -19,7 +19,7 @@ import numpy as np
 from keras.layers import DepthwiseConv2D
 
 from model_compression_toolkit.keras.constants import ACTIVATION, LINEAR
-from tests.common_tests.helpers.generate_test_hw_model import get_quantization_disabled_keras_hw_model
+from tests.keras_tests.fw_hw_model_keras import get_quantization_disabled_keras_hw_model
 from tests.common_tests.helpers.tensors_compare import cosine_similarity
 from tests.keras_tests.feature_networks_tests.base_keras_feature_test import BaseKerasFeatureNetworkTest
 import model_compression_toolkit as mct
