+++ tests/keras_tests/feature_networks_tests/feature_networks/bn_folding_test.py
@@ -22,7 +22,8 @@ import model_compression_toolkit as mct
 import tensorflow as tf
 
 from tests.common_tests.base_layer_test import LayerTestMode
-from tests.common_tests.helpers.generate_test_hw_model import generate_test_hw_model, get_16bit_fw_hw_model
+from tests.common_tests.helpers.generate_test_hw_model import generate_test_hw_model
+from tests.keras_tests.fw_hw_model_keras import get_16bit_fw_hw_model
 from tests.keras_tests.feature_networks_tests.base_keras_feature_test import BaseKerasFeatureNetworkTest
 import numpy as np
 from tests.common_tests.helpers.tensors_compare import cosine_similarity
