+++ tests/keras_tests/feature_networks_tests/feature_networks/multi_head_attention_test.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-from tests.common_tests.helpers.generate_test_hw_model import get_quantization_disabled_keras_hw_model
+from tests.keras_tests.fw_hw_model_keras import get_quantization_disabled_keras_hw_model
 from tests.keras_tests.feature_networks_tests.base_keras_feature_test import BaseKerasFeatureNetworkTest
 from model_compression_toolkit.common.quantization.quantization_config import QuantizationConfig
 import model_compression_toolkit as mct
