+++ tests/keras_tests/feature_networks_tests/feature_networks/multiple_outputs_node_tests.py
@@ -18,10 +18,10 @@ from tests.common_tests.base_feature_test import BaseFeatureNetworkTest
 import model_compression_toolkit as mct
 import tensorflow as tf
 
-from tests.common_tests.helpers.generate_test_hw_model import get_quantization_disabled_keras_hw_model
 from tests.keras_tests.feature_networks_tests.base_keras_feature_test import BaseKerasFeatureNetworkTest
 import numpy as np
 from tests.common_tests.helpers.tensors_compare import cosine_similarity
+from tests.keras_tests.fw_hw_model_keras import get_quantization_disabled_keras_hw_model
 
 keras = tf.keras
 layers = keras.layers
