+++ tests/keras_tests/feature_networks_tests/feature_networks/shift_neg_activation_test.py
@@ -17,9 +17,8 @@ import tensorflow as tf
 
 from model_compression_toolkit.common.constants import SHIFT_NEGATIVE_NON_LINEAR_NUM_BITS
 from model_compression_toolkit.common.network_editors import EditRule, node_filters, actions
-from model_compression_toolkit.hardware_models.keras_hardware_model.keras_default import generate_fhw_model_keras, \
-    get_default_hwm_keras
-from tests.common_tests.helpers.generate_test_hw_model import generate_test_hw_model, get_16bit_fw_hw_model
+from model_compression_toolkit.hardware_models.keras_hardware_model.keras_default import get_default_hwm_keras
+from tests.keras_tests.fw_hw_model_keras import get_16bit_fw_hw_model
 
 if tf.__version__ < "2.6":
     from tensorflow.python.keras.layers.core import TFOpLambda
@@ -27,7 +26,6 @@ else:
     from keras.layers.core import TFOpLambda
 from tests.keras_tests.feature_networks_tests.base_keras_feature_test import BaseKerasFeatureNetworkTest
 import numpy as np
-from tests.common_tests.helpers.tensors_compare import cosine_similarity
 
 keras = tf.keras
 layers = keras.layers
